Make ShaderManager an interface
Add Headless ShaderManager
